fix _bytes_to_bint()

+++ minipg/minipg.pxd
@@ -126,7 +126,7 @@ cdef int PG_TYPE_ANYRANGE
 
 cdef object _decode_column(data, int oid, str encoding)
 
-@cython.locals(r=cython.int)
+@cython.locals(r=cython.longlong)
 cdef long long _bytes_to_bint(bytes b)
 
 cdef bytes _bint_to_bytes(int val)
